add guard clause for delete confirmation

+++ src/Pages/Calendar/Calendar.tsx
@@ -45,6 +45,7 @@ const Calendar = () => {
             })
     }
     async function deleteFamilyMember() {
+        if (!window.confirm(`Are you sure you want to delete ${familyMember.name}? This action cannot be undone!`)) return;
         await fetch(`${API.API_URL}/familyMember/${familyMember.id}`, {
                 method: 'DELETE',
             }).then((res) => {
@@ -162,11 +163,11 @@ const Calendar = () => {
                 <div className="edit-menu">
                     <p>Edit {familyMember.name}</p>
                     <label htmlFor="name-changer">Name</label>
-                    <input type="text" id="name-changer" onChange={event => {
+                    <input type="text" id="name-changer" value={familyMember.name} onChange={event => {
                         setFamilyMember((member) => {
                             return {...member, name: event.target.value}
                         })
-                    }}>{familyMember.name}</input>
+                    }}></input>
 
                     <button className="red-button" onClick={editFamilyMember}>Save changes</button>
                     <button className="red-button" onClick={deleteFamilyMember}>Delete family member</button>
